vision-guard/requirements.txt
flask==3.1.*
ultralytics==8.3.*
opencv-python-headless==4.11.*
Pillow>=10.0
numpy>=1.26
